config.py
"""Configuration — loads secrets from .env file."""

import os
from dotenv import load_dotenv

load_dotenv()

TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN", "")
NOTION_TOKEN = os.getenv("NOTION_TOKEN", "")
NOTION_DATABASE_ID = os.getenv("NOTION_DATABASE_ID", "")

# Telegram bot API limit for file uploads
TELEGRAM_FILE_LIMIT_MB = 50
