requirements.txt
python-telegram-bot>=20.0
yt-dlp
notion-client
python-dotenv
